+++ bonsaidb/examples/server.rs
@@ -70,7 +70,7 @@ async fn main() -> anyhow::Result<()> {
     {
         // To connect over websockets, use the websocket scheme.
         tasks.push(do_some_database_work(
-            Client::<()>::new(Url::parse("ws://localhost:8080")?)
+            Client::new(Url::parse("ws://localhost:8080")?)
                 .await?
                 .database::<Shape>("my-database")
                 .await?,
@@ -80,13 +80,12 @@ async fn main() -> anyhow::Result<()> {
 
     // To connect over QUIC, use the bonsaidb scheme.
     tasks.push(do_some_database_work(
-        Client::<()>::new_with_certificate(
-            Url::parse("bonsaidb://localhost")?,
-            Some(certificate),
-        )
-        .await?
-        .database::<Shape>("my-database")
-        .await?,
+        Client::<()>::build(Url::parse("bonsaidb://localhost")?)
+            .with_certificate(certificate)
+            .finish()
+            .await?
+            .database::<Shape>("my-database")
+            .await?,
         "bonsaidb",
     ));